// itsm-cli/src/lib/types.ts
export interface Notification {
  id: number; title: string; content?: string;
  type: string; read: boolean; level?: string;
  created_at: string;
}